Filtered students; Added checkin_status
